+++ spacy/util.py
@@ -20,7 +20,6 @@ import subprocess
 from contextlib import contextmanager
 import tempfile
 import shutil
-import hashlib
 import shlex
 
 try:
@@ -509,34 +508,6 @@ def make_tempdir():
         shutil.rmtree(str(d))
     except PermissionError as e:
         warnings.warn(Warnings.W091.format(dir=d, msg=e))
-
-
-def get_hash(data) -> str:
-    """Get the hash for a JSON-serializable object.
-
-    data: The data to hash.
-    RETURNS (str): The hash.
-    """
-    data_str = srsly.json_dumps(data, sort_keys=True).encode("utf8")
-    return hashlib.md5(data_str).hexdigest()
-
-
-def get_checksum(path: Union[Path, str]) -> str:
-    """Get the checksum for a file or directory given its file path. If a
-    directory path is provided, this uses all files in that directory.
-
-    path (Union[Path, str]): The file or directory path.
-    RETURNS (str): The checksum.
-    """
-    path = Path(path)
-    if path.is_file():
-        return hashlib.md5(Path(path).read_bytes()).hexdigest()
-    if path.is_dir():
-        dir_checksum = hashlib.md5()
-        for sub_file in sorted(fp for fp in path.rglob("*") if fp.is_file()):
-            dir_checksum.update(sub_file.read_bytes())
-        return dir_checksum.hexdigest()
-    raise ValueError(Errors.E968.format(path=path))
 
 
 def is_cwd(path: Union[Path, str]) -> bool:
